moved encounter resource to resource folder
